requirements.txt
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
elevenlabs>=1.0.0
